Refactor: Move constants to proper file
Refactor: Move constants to proper file
